src/hornbill/pyproject.toml
[project]
name = "actual"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.120.4",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
